Renamed folder from surface_intersect to surface_query
