+++ backend_go/surface_query/utils/blob_fetcher.go
@@ -2,7 +2,7 @@ package utils
 
 import (
 	"fmt"
-	"surface_intersect/xtgeo"
+	"surface_query/xtgeo"
 )
 
 type BlobFetcher struct {
