+++ backend_go/surface_query/xtgeo/blob_to_surface.go
