+++ backend_go/surface_query/xtgeo/calculate_vector_length_and_angle.go
