+++ backend_go/surface_query/xtgeo/constants.go
