+++ backend_go/surface_query/xtgeo/distance_between_point_and_line.go
