+++ backend_go/surface_query/xtgeo/structs.go
