+++ backend_go/surface_query/xtgeo/surface_z_from_xy.go
